bugfix: Print copying error together with stacktrace
It seems we don't get that properly forwarded to Bloop otherwise.
