+++ backend/src/main/scala/bloop/logging/BloopLogger.scala
@@ -146,13 +146,6 @@ object BloopLogger {
   def default(name: String): BloopLogger =
     at(name, System.out, System.err, false, DebugFilter.All)
 
-  def prettyPrintException(t: Throwable): String = {
-    val sw = new java.io.StringWriter()
-    val pw = new java.io.PrintWriter(sw)
-    t.printStackTrace(pw)
-    sw.toString()
-  }
-
   private lazy val colorsRegex = "\u001b\\[[0-9;]*m".r
 
   /**
